DocQA/requirements.txt
flask==3.0.0
werkzeug==3.0.1
python-dotenv==1.0.0
openai==1.12.0
anthropic==0.18.1
chromadb==0.4.22
sentence-transformers==2.3.1
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.3
lxml==5.1.0
Pillow==10.2.0
pytesseract==0.3.10
